{273ae71f-305a-49ba-epoc-173603972731}

+++ widgets/python/_rightThumb/_vars/__init__.py
@@ -1478,8 +1478,10 @@ def get_administrator_sid():
 		temp_file_name = temp_file.name
 
 	try:
-		# os.system(f'wmic useraccount where "name=\'administrator\' and domain=\'%computername%\'" get name,sid | find "admin" > {temp_file_name}')
-		os.system(f'powershell -Command "Get-WmiObject Win32_UserAccount | Where-Object {{ $_.Name -eq \'administrator\' }} | Select-Object Name,SID | Out-File -FilePath {temp_file_name}"')
+		if os.path.isfile('C:\Windows\System32\wbem\WMIC.exe'):
+			os.system(f'wmic useraccount where "name=\'administrator\' and domain=\'%computername%\'" get name,sid | find "admin" > {temp_file_name}')
+		else:
+			os.system(f'powershell -Command "Get-WmiObject Win32_UserAccount | Where-Object {{ $_.Name -eq \'administrator\' }} | Select-Object Name,SID | Out-File -FilePath {temp_file_name}"')
 		# try:
 		# except:
